# CUDA: stream-k decomposition for MMQ