This PR implements "stream-k decomposition" as described in [this](https://arxiv.org/abs/2301.03598) paper for the MMQ kernels. The idea is that instead of dividing the output tensor into tiles and assigning those tiles to the streaming multiprocessors in waves you instead assign partial tiles to the streaming multiprocessors that you then combine with an additional "fixup" kernel. Notably the number of tiles that you need to fix is constant and equal to the number of streaming multiprocessors (vs. e.g. splitting all tiles).

<details>
<summary>Performance vs. master MMQ</summary>

| GPU        | Model             |     Microbatch size | Test     |     t/s master |     t/s cuda-mmq-stream-k-2 |     Speedup |
| :--------- | :---------------- | ------------------: | :------- | -------------: | --------------------------: | ----------: |
| RTX 4090   | llama 8B Q2_K_M   |                  16 | pp2048   |        1845.03 |                     1974.88 |        1.07 |
| RTX 4090   | llama 8B Q2_K_M   |                  32 | pp2048   |        2704.27 |                     3052.64 |        1.13 |
| RTX 4090   | llama 8B Q2_K_M   |                  63 | pp2048   |        3974.88 |                     4310.72 |        1.08 |
| RTX 4090   | llama 8B Q2_K_M   |                 128 | pp2048   |        5464.22 |                     5528.70 |        1.01 |
| RTX 4090   | llama 8B Q2_K_M   |                 256 | pp2048   |        6815.42 |                     7109.27 |        1.04 |
| RTX 4090   | llama 8B Q2_K_M   |                 512 | pp2048   |        7416.65 |                     7558.46 |        1.02 |
| RTX 4090   | llama 8B Q2_K_M   |                1024 | pp2048   |        7537.06 |                     7469.47 |        0.99 |
| RTX 4090   | llama 8B Q2_K_M   |                2048 | pp2048   |        7004.78 |                     6950.96 |        0.99 |
| RTX 4090   | llama 8B Q3_K_S   |                  16 | pp2048   |        1657.97 |                     1766.89 |        1.07 |
| RTX 4090   | llama 8B Q3_K_S   |                  32 | pp2048   |        2615.77 |                     2986.33 |        1.14 |
| RTX 4090   | llama 8B Q3_K_S   |                  63 | pp2048   |        4055.24 |                     4484.16 |        1.11 |
| RTX 4090   | llama 8B Q3_K_S   |                 128 | pp2048   |        5920.58 |                     6369.46 |        1.08 |
| RTX 4090   | llama 8B Q3_K_S   |                 256 | pp2048   |        7733.07 |                     8454.78 |        1.09 |
| RTX 4090   | llama 8B Q3_K_S   |                 512 | pp2048   |        8676.65 |                     8966.30 |        1.03 |
| RTX 4090   | llama 8B Q3_K_S   |                1024 | pp2048   |        8708.10 |                     8828.17 |        1.01 |
| RTX 4090   | llama 8B Q3_K_S   |                2048 | pp2048   |        8053.46 |                     8111.16 |        1.01 |
| RTX 4090   | llama 8B Q4_0     |                  16 | pp2048   |        1873.62 |                     1989.96 |        1.06 |
| RTX 4090   | llama 8B Q4_0     |                  32 | pp2048   |        3269.12 |                     3456.99 |        1.06 |
| RTX 4090   | llama 8B Q4_0     |                  63 | pp2048   |        5104.59 |                     5284.82 |        1.04 |
| RTX 4090   | llama 8B Q4_0     |                 128 | pp2048   |        7114.91 |                     7227.59 |        1.02 |
| RTX 4090   | llama 8B Q4_0     |                 256 | pp2048   |        9318.58 |                     9514.91 |        1.02 |
| RTX 4090   | llama 8B Q4_0     |                 512 | pp2048   |       10198.79 |                    10317.21 |        1.01 |
| RTX 4090   | llama 8B Q4_0     |                1024 | pp2048   |       10158.53 |                    10065.18 |        0.99 |
| RTX 4090   | llama 8B Q4_0     |                2048 | pp2048   |        9241.77 |                     9119.39 |        0.99 |
| RTX 4090   | llama 8B Q4_1     |                  16 | pp2048   |        1858.86 |                     1880.17 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                  32 | pp2048   |        3179.30 |                     2999.65 |        0.94 |
| RTX 4090   | llama 8B Q4_1     |                  63 | pp2048   |        5412.60 |                     5084.00 |        0.94 |
| RTX 4090   | llama 8B Q4_1     |                 128 | pp2048   |        6581.02 |                     6982.92 |        1.06 |
| RTX 4090   | llama 8B Q4_1     |                 256 | pp2048   |        8990.72 |                     9070.09 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                 512 | pp2048   |        9667.64 |                     9783.97 |        1.01 |
| RTX 4090   | llama 8B Q4_1     |                1024 | pp2048   |        9930.18 |                     9568.42 |        0.96 |
| RTX 4090   | llama 8B Q4_1     |                2048 | pp2048   |        9154.70 |                     8782.85 |        0.96 |
| RTX 4090   | llama 8B Q4_K_S   |                  16 | pp2048   |        1937.52 |                     2004.38 |        1.03 |
| RTX 4090   | llama 8B Q4_K_S   |                  32 | pp2048   |        3342.11 |                     3399.52 |        1.02 |
| RTX 4090   | llama 8B Q4_K_S   |                  63 | pp2048   |        5154.91 |                     5080.20 |        0.99 |
| RTX 4090   | llama 8B Q4_K_S   |                 128 | pp2048   |        6874.51 |                     6758.06 |        0.98 |
| RTX 4090   | llama 8B Q4_K_S   |                 256 | pp2048   |        8434.04 |                     8750.48 |        1.04 |
| RTX 4090   | llama 8B Q4_K_S   |                 512 | pp2048   |        9195.77 |                     9421.38 |        1.02 |
| RTX 4090   | llama 8B Q4_K_S   |                1024 | pp2048   |        9169.23 |                     9306.70 |        1.01 |
| RTX 4090   | llama 8B Q4_K_S   |                2048 | pp2048   |        8629.07 |                     8523.98 |        0.99 |
| RTX 4090   | llama 8B Q5_0     |                  16 | pp2048   |        1480.94 |                     1673.97 |        1.13 |
| RTX 4090   | llama 8B Q5_0     |                  32 | pp2048   |        2509.59 |                     3051.97 |        1.22 |
| RTX 4090   | llama 8B Q5_0     |                  63 | pp2048   |        4162.96 |                     4628.07 |        1.11 |
| RTX 4090   | llama 8B Q5_0     |                 128 | pp2048   |        6207.95 |                     6629.96 |        1.07 |
| RTX 4090   | llama 8B Q5_0     |                 256 | pp2048   |        8006.74 |                     8867.56 |        1.11 |
| RTX 4090   | llama 8B Q5_0     |                 512 | pp2048   |        9223.85 |                     9702.41 |        1.05 |
| RTX 4090   | llama 8B Q5_0     |                1024 | pp2048   |        9277.39 |                     9535.97 |        1.03 |
| RTX 4090   | llama 8B Q5_0     |                2048 | pp2048   |        8710.12 |                     8780.70 |        1.01 |
| RTX 4090   | llama 8B Q5_1     |                  16 | pp2048   |        1619.67 |                     1580.42 |        0.98 |
| RTX 4090   | llama 8B Q5_1     |                  32 | pp2048   |        2584.57 |                     2547.53 |        0.99 |
| RTX 4090   | llama 8B Q5_1     |                  63 | pp2048   |        4297.61 |                     4222.33 |        0.98 |
| RTX 4090   | llama 8B Q5_1     |                 128 | pp2048   |        5975.51 |                     6523.30 |        1.09 |
| RTX 4090   | llama 8B Q5_1     |                 256 | pp2048   |        7744.80 |                     8451.57 |        1.09 |
| RTX 4090   | llama 8B Q5_1     |                 512 | pp2048   |        8823.13 |                     9127.40 |        1.03 |
| RTX 4090   | llama 8B Q5_1     |                1024 | pp2048   |        8931.16 |                     9039.22 |        1.01 |
| RTX 4090   | llama 8B Q5_1     |                2048 | pp2048   |        8335.92 |                     8360.64 |        1.00 |
| RTX 4090   | llama 8B Q5_K_S   |                  16 | pp2048   |        1669.68 |                     1705.77 |        1.02 |
| RTX 4090   | llama 8B Q5_K_S   |                  32 | pp2048   |        2776.29 |                     2777.08 |        1.00 |
| RTX 4090   | llama 8B Q5_K_S   |                  63 | pp2048   |        4419.73 |                     4356.69 |        0.99 |
| RTX 4090   | llama 8B Q5_K_S   |                 128 | pp2048   |        5981.22 |                     6426.45 |        1.07 |
| RTX 4090   | llama 8B Q5_K_S   |                 256 | pp2048   |        7698.50 |                     8275.37 |        1.07 |
| RTX 4090   | llama 8B Q5_K_S   |                 512 | pp2048   |        8615.20 |                     8939.98 |        1.04 |
| RTX 4090   | llama 8B Q5_K_S   |                1024 | pp2048   |        8703.88 |                     8884.14 |        1.02 |
| RTX 4090   | llama 8B Q5_K_S   |                2048 | pp2048   |        8158.43 |                     8242.38 |        1.01 |
| RTX 4090   | llama 8B Q6_K     |                  16 | pp2048   |        1410.39 |                     1455.31 |        1.03 |
| RTX 4090   | llama 8B Q6_K     |                  32 | pp2048   |        2452.60 |                     2754.07 |        1.12 |
| RTX 4090   | llama 8B Q6_K     |                  63 | pp2048   |        4258.11 |                     4430.69 |        1.04 |
| RTX 4090   | llama 8B Q6_K     |                 128 | pp2048   |        6049.71 |                     6309.51 |        1.04 |
| RTX 4090   | llama 8B Q6_K     |                 256 | pp2048   |        7968.27 |                     8277.67 |        1.04 |
| RTX 4090   | llama 8B Q6_K     |                 512 | pp2048   |        8751.79 |                     8968.23 |        1.02 |
| RTX 4090   | llama 8B Q6_K     |                1024 | pp2048   |        8837.98 |                     8852.53 |        1.00 |
| RTX 4090   | llama 8B Q6_K     |                2048 | pp2048   |        8155.85 |                     8044.97 |        0.99 |
| RTX 4090   | llama 8B Q8_0     |                  16 | pp2048   |        1123.00 |                     1276.92 |        1.14 |
| RTX 4090   | llama 8B Q8_0     |                  32 | pp2048   |        2178.68 |                     2374.20 |        1.09 |
| RTX 4090   | llama 8B Q8_0     |                  63 | pp2048   |        3911.79 |                     4065.28 |        1.04 |
| RTX 4090   | llama 8B Q8_0     |                 128 | pp2048   |        6343.39 |                     6563.01 |        1.03 |
| RTX 4090   | llama 8B Q8_0     |                 256 | pp2048   |        8656.65 |                     9411.70 |        1.09 |
| RTX 4090   | llama 8B Q8_0     |                 512 | pp2048   |        9933.38 |                    10591.37 |        1.07 |
| RTX 4090   | llama 8B Q8_0     |                1024 | pp2048   |       10310.62 |                    10554.57 |        1.02 |
| RTX 4090   | llama 8B Q8_0     |                2048 | pp2048   |        9672.82 |                     9544.11 |        0.99 |
| RTX 3090   | llama 8B Q2_K_M   |                  16 | pp2048   |         860.08 |                      983.08 |        1.14 |
| RTX 3090   | llama 8B Q2_K_M   |                  32 | pp2048   |        1191.51 |                     1425.70 |        1.20 |
| RTX 3090   | llama 8B Q2_K_M   |                  63 | pp2048   |        1575.38 |                     1893.31 |        1.20 |
| RTX 3090   | llama 8B Q2_K_M   |                 128 | pp2048   |        2007.22 |                     2262.78 |        1.13 |
| RTX 3090   | llama 8B Q2_K_M   |                 256 | pp2048   |        2487.20 |                     2536.22 |        1.02 |
| RTX 3090   | llama 8B Q2_K_M   |                 512 | pp2048   |        2592.80 |                     2647.31 |        1.02 |
| RTX 3090   | llama 8B Q2_K_M   |                1024 | pp2048   |        2646.01 |                     2639.46 |        1.00 |
| RTX 3090   | llama 8B Q2_K_M   |                2048 | pp2048   |        2604.77 |                     2608.64 |        1.00 |
| RTX 3090   | llama 8B Q3_K_S   |                  16 | pp2048   |         809.98 |                      920.54 |        1.14 |
| RTX 3090   | llama 8B Q3_K_S   |                  32 | pp2048   |        1139.97 |                     1485.39 |        1.30 |
| RTX 3090   | llama 8B Q3_K_S   |                  63 | pp2048   |        1644.27 |                     2087.77 |        1.27 |
| RTX 3090   | llama 8B Q3_K_S   |                 128 | pp2048   |        2331.84 |                     2736.05 |        1.17 |
| RTX 3090   | llama 8B Q3_K_S   |                 256 | pp2048   |        2906.75 |                     3057.93 |        1.05 |
| RTX 3090   | llama 8B Q3_K_S   |                 512 | pp2048   |        3053.04 |                     3191.71 |        1.05 |
| RTX 3090   | llama 8B Q3_K_S   |                1024 | pp2048   |        3173.62 |                     3236.57 |        1.02 |
| RTX 3090   | llama 8B Q3_K_S   |                2048 | pp2048   |        3092.63 |                     3172.58 |        1.03 |
| RTX 3090   | llama 8B Q4_0     |                  16 | pp2048   |        1076.43 |                     1173.33 |        1.09 |
| RTX 3090   | llama 8B Q4_0     |                  32 | pp2048   |        1556.26 |                     1788.35 |        1.15 |
| RTX 3090   | llama 8B Q4_0     |                  63 | pp2048   |        2144.28 |                     2534.95 |        1.18 |
| RTX 3090   | llama 8B Q4_0     |                 128 | pp2048   |        2830.76 |                     3166.94 |        1.12 |
| RTX 3090   | llama 8B Q4_0     |                 256 | pp2048   |        3438.60 |                     3599.03 |        1.05 |
| RTX 3090   | llama 8B Q4_0     |                 512 | pp2048   |        3701.21 |                     3784.69 |        1.02 |
| RTX 3090   | llama 8B Q4_0     |                1024 | pp2048   |        3832.97 |                     3795.66 |        0.99 |
| RTX 3090   | llama 8B Q4_0     |                2048 | pp2048   |        3643.18 |                     3634.57 |        1.00 |
| RTX 3090   | llama 8B Q4_1     |                  16 | pp2048   |        1179.37 |                     1268.37 |        1.08 |
| RTX 3090   | llama 8B Q4_1     |                  32 | pp2048   |        1614.90 |                     1652.84 |        1.02 |
| RTX 3090   | llama 8B Q4_1     |                  63 | pp2048   |        2031.74 |                     2382.21 |        1.17 |
| RTX 3090   | llama 8B Q4_1     |                 128 | pp2048   |        2755.17 |                     2966.05 |        1.08 |
| RTX 3090   | llama 8B Q4_1     |                 256 | pp2048   |        3232.95 |                     3322.46 |        1.03 |
| RTX 3090   | llama 8B Q4_1     |                 512 | pp2048   |        3579.61 |                     3469.01 |        0.97 |
| RTX 3090   | llama 8B Q4_1     |                1024 | pp2048   |        3718.61 |                     3526.51 |        0.95 |
| RTX 3090   | llama 8B Q4_1     |                2048 | pp2048   |        3624.65 |                     3442.58 |        0.95 |
| RTX 3090   | llama 8B Q4_K_S   |                  16 | pp2048   |        1125.44 |                     1242.52 |        1.10 |
| RTX 3090   | llama 8B Q4_K_S   |                  32 | pp2048   |        1585.72 |                     1693.02 |        1.07 |
| RTX 3090   | llama 8B Q4_K_S   |                  63 | pp2048   |        2032.91 |                     2312.02 |        1.14 |
| RTX 3090   | llama 8B Q4_K_S   |                 128 | pp2048   |        2564.13 |                     2812.79 |        1.10 |
| RTX 3090   | llama 8B Q4_K_S   |                 256 | pp2048   |        3073.94 |                     3181.44 |        1.03 |
| RTX 3090   | llama 8B Q4_K_S   |                 512 | pp2048   |        3313.49 |                     3325.57 |        1.00 |
| RTX 3090   | llama 8B Q4_K_S   |                1024 | pp2048   |        3444.24 |                     3376.12 |        0.98 |
| RTX 3090   | llama 8B Q4_K_S   |                2048 | pp2048   |        3250.36 |                     3291.11 |        1.01 |
| RTX 3090   | llama 8B Q5_0     |                  16 | pp2048   |         776.15 |                      946.39 |        1.22 |
| RTX 3090   | llama 8B Q5_0     |                  32 | pp2048   |        1217.41 |                     1621.45 |        1.33 |
| RTX 3090   | llama 8B Q5_0     |                  63 | pp2048   |        1687.45 |                     2198.64 |        1.30 |
| RTX 3090   | llama 8B Q5_0     |                 128 | pp2048   |        2404.10 |                     2902.89 |        1.21 |
| RTX 3090   | llama 8B Q5_0     |                 256 | pp2048   |        3087.93 |                     3296.47 |        1.07 |
| RTX 3090   | llama 8B Q5_0     |                 512 | pp2048   |        3219.13 |                     3451.26 |        1.07 |
| RTX 3090   | llama 8B Q5_0     |                1024 | pp2048   |        3363.65 |                     3484.96 |        1.04 |
| RTX 3090   | llama 8B Q5_0     |                2048 | pp2048   |        3298.89 |                     3401.04 |        1.03 |
| RTX 3090   | llama 8B Q5_1     |                  16 | pp2048   |         930.03 |                      987.19 |        1.06 |
| RTX 3090   | llama 8B Q5_1     |                  32 | pp2048   |        1365.85 |                     1423.36 |        1.04 |
| RTX 3090   | llama 8B Q5_1     |                  63 | pp2048   |        1727.76 |                     2091.52 |        1.21 |
| RTX 3090   | llama 8B Q5_1     |                 128 | pp2048   |        2362.19 |                     2726.86 |        1.15 |
| RTX 3090   | llama 8B Q5_1     |                 256 | pp2048   |        2978.09 |                     3068.63 |        1.03 |
| RTX 3090   | llama 8B Q5_1     |                 512 | pp2048   |        3090.98 |                     3242.92 |        1.05 |
| RTX 3090   | llama 8B Q5_1     |                1024 | pp2048   |        3216.80 |                     3274.03 |        1.02 |
| RTX 3090   | llama 8B Q5_1     |                2048 | pp2048   |        3187.65 |                     3211.42 |        1.01 |
| RTX 3090   | llama 8B Q5_K_S   |                  16 | pp2048   |         903.89 |                     1016.46 |        1.12 |
| RTX 3090   | llama 8B Q5_K_S   |                  32 | pp2048   |        1296.51 |                     1448.98 |        1.12 |
| RTX 3090   | llama 8B Q5_K_S   |                  63 | pp2048   |        1715.60 |                     2032.78 |        1.18 |
| RTX 3090   | llama 8B Q5_K_S   |                 128 | pp2048   |        2326.05 |                     2652.85 |        1.14 |
| RTX 3090   | llama 8B Q5_K_S   |                 256 | pp2048   |        2894.26 |                     3004.17 |        1.04 |
| RTX 3090   | llama 8B Q5_K_S   |                 512 | pp2048   |        3061.19 |                     3129.73 |        1.02 |
| RTX 3090   | llama 8B Q5_K_S   |                1024 | pp2048   |        3163.91 |                     3190.35 |        1.01 |
| RTX 3090   | llama 8B Q5_K_S   |                2048 | pp2048   |        3097.12 |                     3135.12 |        1.01 |
| RTX 3090   | llama 8B Q6_K     |                  16 | pp2048   |         813.79 |                      926.20 |        1.14 |
| RTX 3090   | llama 8B Q6_K     |                  32 | pp2048   |        1251.80 |                     1552.21 |        1.24 |
| RTX 3090   | llama 8B Q6_K     |                  63 | pp2048   |        1771.66 |                     2216.34 |        1.25 |
…