@ggerganov 2024-06-20
Here are the results on the RTX 2060, data for 1B and 7B models:

```bash
LLAMA_CUDA=1 ./scripts/compare-commits.sh master da1db13d6aaabd0889a1bc1c7755a462621bd1d6 \
  -m models/llama-7b-v2/ggml-model-q4_0.gguf \
  -m ./models/llama-7b-v2/ggml-model-q4_k.gguf \
  -m models/tinyllama-1b/ggml-model-q4_0.gguf \
  -m models/tinyllama-1b/ggml-model-q8_0.gguf \
  -p 2048 -ub 16,32,64,128,256,512,1024,2048 -ngl 99 -n 0
```

<details>

<summary>Performance vs. master cuBLAS</summary>

ggml_cuda_init: GGML_CUDA_FORCE_MMQ:   no
ggml_cuda_init: CUDA_USE_TENSOR_CORES: yes
ggml_cuda_init: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 2060 SUPER, compute capability 7.5, VMM: yes

| Model           |   [GiB] |  Microbatch size | Test   |   t/s master |   t/s da1db13d |   Speedup |
|:----------------|--------:|-----------------:|:-------|-------------:|---------------:|----------:|
| llama 1B Q4_0   |    0.59 |               16 | pp2048 |      1181.60 |        1146.10 |      0.97 |
| llama 1B Q4_0   |    0.59 |               32 | pp2048 |      2544.40 |        2376.44 |      0.93 |
| llama 1B Q4_0   |    0.59 |               64 | pp2048 |      3302.75 |        3231.90 |      0.98 |
| llama 1B Q4_0   |    0.59 |              128 | pp2048 |      3330.94 |        3327.80 |      1.00 |
| llama 1B Q4_0   |    0.59 |              256 | pp2048 |      4179.32 |        4177.13 |      1.00 |
| llama 1B Q4_0   |    0.59 |              512 | pp2048 |      4597.28 |        4595.68 |      1.00 |
| llama 1B Q4_0   |    0.59 |             1024 | pp2048 |      4587.57 |        4581.55 |      1.00 |
| llama 1B Q4_0   |    0.59 |             2048 | pp2048 |      4218.85 |        4211.63 |      1.00 |
| llama 1B Q8_0   |    1.09 |               16 | pp2048 |       957.18 |        1028.21 |      1.07 |
| llama 1B Q8_0   |    1.09 |               32 | pp2048 |      2055.63 |        2281.86 |      1.11 |
| llama 1B Q8_0   |    1.09 |               64 | pp2048 |      2814.76 |        3077.22 |      1.09 |
| llama 1B Q8_0   |    1.09 |              128 | pp2048 |      3662.31 |        3659.32 |      1.00 |
| llama 1B Q8_0   |    1.09 |              256 | pp2048 |      4428.42 |        4427.45 |      1.00 |
| llama 1B Q8_0   |    1.09 |              512 | pp2048 |      4745.80 |        4746.90 |      1.00 |
| llama 1B Q8_0   |    1.09 |             1024 | pp2048 |      4665.55 |        4660.40 |      1.00 |
| llama 1B Q8_0   |    1.09 |             2048 | pp2048 |      4255.26 |        4247.57 |      1.00 |
| llama 7B Q4_0   |    3.56 |               16 | pp2048 |       429.42 |         405.31 |      0.94 |
| llama 7B Q4_0   |    3.56 |               32 | pp2048 |       707.97 |         675.67 |      0.95 |
| llama 7B Q4_0   |    3.56 |               64 | pp2048 |       962.51 |         935.46 |      0.97 |
| llama 7B Q4_0   |    3.56 |              128 | pp2048 |       797.75 |         797.07 |      1.00 |
| llama 7B Q4_0   |    3.56 |              256 | pp2048 |      1097.33 |        1096.56 |      1.00 |
| llama 7B Q4_0   |    3.56 |              512 | pp2048 |      1330.73 |        1329.57 |      1.00 |
| llama 7B Q4_0   |    3.56 |             1024 | pp2048 |      1440.01 |        1438.79 |      1.00 |
| llama 7B Q4_0   |    3.56 |             2048 | pp2048 |      1458.11 |        1456.14 |      1.00 |
| llama 7B Q4_K_M |    3.80 |               16 | pp2048 |       414.12 |         388.30 |      0.94 |
| llama 7B Q4_K_M |    3.80 |               32 | pp2048 |       716.33 |         726.15 |      1.01 |
| llama 7B Q4_K_M |    3.80 |               64 | pp2048 |       952.33 |         899.34 |      0.94 |
| llama 7B Q4_K_M |    3.80 |              128 | pp2048 |       803.76 |         803.57 |      1.00 |
| llama 7B Q4_K_M |    3.80 |              256 | pp2048 |      1103.59 |        1102.60 |      1.00 |
| llama 7B Q4_K_M |    3.80 |              512 | pp2048 |      1334.25 |        1333.16 |      1.00 |
| llama 7B Q4_K_M |    3.80 |             1024 | pp2048 |      1442.62 |        1440.66 |      1.00 |
| llama 7B Q4_K_M |    3.80 |             2048 | pp2048 |      1458.86 |        1457.32 |      1.00 |

</details>

```bash
LLAMA_CUDA_FORCE_MMQ=1 LLAMA_CUDA=1 ./scripts/compare-commits.sh master da1db13d6aaabd0889a1bc1c7755a462621bd1d6 \
  -m models/llama-7b-v2/ggml-model-q4_0.gguf \
  -m ./models/llama-7b-v2/ggml-model-q4_k.gguf \
  -m models/tinyllama-1b/ggml-model-q4_0.gguf \
  -m models/tinyllama-1b/ggml-model-q8_0.gguf \
  -p 2048 -ub 16,32,64,128,256,512,1024,2048 -ngl 99 -n 0
```

<details>

<summary>Performance vs. master MMQ</summary>

ggml_cuda_init: GGML_CUDA_FORCE_MMQ:   yes
ggml_cuda_init: CUDA_USE_TENSOR_CORES: no
ggml_cuda_init: found 1 CUDA devices:
  Device 0: NVIDIA GeForce RTX 2060 SUPER, compute capability 7.5, VMM: yes

| Model           |   [GiB] |  Microbatch size | Test   |   t/s master |   t/s da1db13d |   Speedup |
|:----------------|--------:|-----------------:|:-------|-------------:|---------------:|----------:|
| llama 1B Q4_0   |    0.59 |               16 | pp2048 |      1182.36 |        1145.77 |      0.97 |
| llama 1B Q4_0   |    0.59 |               32 | pp2048 |      2543.57 |        2364.53 |      0.93 |
| llama 1B Q4_0   |    0.59 |               64 | pp2048 |      3300.52 |        3229.23 |      0.98 |
| llama 1B Q4_0   |    0.59 |              128 | pp2048 |      3824.02 |        3710.79 |      0.97 |
| llama 1B Q4_0   |    0.59 |              256 | pp2048 |      4388.32 |        4247.36 |      0.97 |
| llama 1B Q4_0   |    0.59 |              512 | pp2048 |      4280.59 |        4312.24 |      1.01 |
| llama 1B Q4_0   |    0.59 |             1024 | pp2048 |      4108.44 |        4111.93 |      1.00 |
| llama 1B Q4_0   |    0.59 |             2048 | pp2048 |      3755.14 |        3735.72 |      0.99 |
| llama 1B Q8_0   |    1.09 |               16 | pp2048 |       956.76 |        1027.92 |      1.07 |
| llama 1B Q8_0   |    1.09 |               32 | pp2048 |      2054.94 |        2278.54 |      1.11 |
| llama 1B Q8_0   |    1.09 |               64 | pp2048 |      2814.72 |        3077.43 |      1.09 |
| llama 1B Q8_0   |    1.09 |              128 | pp2048 |      3568.13 |        3609.68 |      1.01 |
| llama 1B Q8_0   |    1.09 |              256 | pp2048 |      4169.34 |        4041.25 |      0.97 |
| llama 1B Q8_0   |    1.09 |              512 | pp2048 |      4127.56 |        4104.11 |      0.99 |
| llama 1B Q8_0   |    1.09 |             1024 | pp2048 |      3989.79 |        3912.26 |      0.98 |
| llama 1B Q8_0   |    1.09 |             2048 | pp2048 |      3663.29 |        3574.50 |      0.98 |
| llama 7B Q4_0   |    3.56 |               16 | pp2048 |       429.39 |         405.35 |      0.94 |
| llama 7B Q4_0   |    3.56 |               32 | pp2048 |       707.94 |         676.26 |      0.96 |
| llama 7B Q4_0   |    3.56 |               64 | pp2048 |       962.28 |         936.64 |      0.97 |
| llama 7B Q4_0   |    3.56 |              128 | pp2048 |      1133.00 |        1139.91 |      1.01 |
| llama 7B Q4_0   |    3.56 |              256 | pp2048 |      1135.68 |        1179.10 |      1.04 |
| llama 7B Q4_0   |    3.56 |              512 | pp2048 |      1164.37 |        1188.67 |      1.02 |
| llama 7B Q4_0   |    3.56 |             1024 | pp2048 |      1153.37 |        1164.52 |      1.01 |
| llama 7B Q4_0   |    3.56 |             2048 | pp2048 |      1112.35 |        1117.58 |      1.00 |
| llama 7B Q4_K_M |    3.80 |               16 | pp2048 |       414.09 |         388.46 |      0.94 |
| llama 7B Q4_K_M |    3.80 |               32 | pp2048 |       716.26 |         726.54 |      1.01 |
| llama 7B Q4_K_M |    3.80 |               64 | pp2048 |       952.00 |         900.44 |      0.95 |
| llama 7B Q4_K_M |    3.80 |              128 | pp2048 |      1072.28 |         976.05 |      0.91 |
| llama 7B Q4_K_M |    3.80 |              256 | pp2048 |      1068.74 |        1076.58 |      1.01 |
| llama 7B Q4_K_M |    3.80 |              512 | pp2048 |      1113.67 |        1106.75 |      0.99 |
| llama 7B Q4_K_M |    3.80 |             1024 | pp2048 |      1105.77 |        1104.62 |      1.00 |
| llama 7B Q4_K_M |    3.80 |             2048 | pp2048 |      1077.17 |        1072.15 |      1.00 |

</details>