@slaren 2024-06-20
Anything that improves the testing capabilities would be good. I plan to do some work on `test-backend-ops` to allow testing random values for the parameters of the ops (fuzzing), that should make it easier to find edge cases in ops.

I tested the performance with the 3090 Ti and it is similar to your results with the 3090 as usual. I also tried with a 3080 and it is not always an improvement. However this GPU is my display device and the results are not as repeatable as with the 3090 Ti.

| GPU      | Model         |   Model Size [GiB] |   Microbatch size | Test   |   t/s master |   t/s cuda-mmq-stream-k-2 |   Speedup |
|:---------|:--------------|-------------------:|------------------:|:-------|-------------:|--------------------------:|----------:|
| RTX 3080 | llama 7B Q4_0 |               3.56 |                16 | pp2048 |       892.93 |                    817.34 |      0.92 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |                32 | pp2048 |      1545.34 |                   1502.87 |      0.97 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |                64 | pp2048 |      2103.63 |                   2111.72 |      1.00 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |               128 | pp2048 |      2679.05 |                   2657.88 |      0.99 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |               256 | pp2048 |      3159.96 |                   2935.99 |      0.93 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |               512 | pp2048 |      3144.34 |                   3035.00 |      0.97 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |              1024 | pp2048 |      3121.95 |                   2937.57 |      0.94 |
| RTX 3080 | llama 7B Q4_0 |               3.56 |              2048 | pp2048 |      2986.47 |                   2825.06 |      0.95 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |                16 | pp2048 |       675.32 |                    697.10 |      1.03 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |                32 | pp2048 |      1312.00 |                   1368.82 |      1.04 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |                64 | pp2048 |      1931.47 |                   2065.59 |      1.07 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |               128 | pp2048 |      2596.44 |                   2714.64 |      1.05 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |               256 | pp2048 |      3139.19 |                   2995.61 |      0.95 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |               512 | pp2048 |      3149.73 |                   3132.67 |      0.99 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |              1024 | pp2048 |      3137.40 |                   3086.56 |      0.98 |
| RTX 3080 | llama 7B Q8_0 |               6.67 |              2048 | pp2048 |      3016.30 |                   2938.36 |      0.97 |